BCR-95: handled empty catalog search
